requirements.txt
python-telegram-bot==20.7
langchain==0.1.0
langchain-openai==0.1.0
langchain-community==0.0.10
python-dotenv==1.0.0
requests==2.31.0